Add project layer separation
The project was built with only one layer (API layer) and divided by folders (I thought i could only do this with Visual Studio IDE, but then i found the dotnet cli commands to create classlibs and solutions). This was fixed by creating a Core layer (classlib) for Services, interfaces and entities. A new solution was created on the root folder (/fullstack-challenge), and I added the Core and API classlibs to it. This commit is huge, but it was necessary and it was explained by this commit description.